Add TextTool

+++ LiveShot.API/Drawing/DrawingTool.cs
@@ -14,7 +14,7 @@ namespace LiveShot.API.Drawing
         public abstract void OnMouseLeftButtonUp(MouseButtonEventArgs e, AbstractDrawCanvas canvas);
         public abstract void OnMouseMove(MouseEventArgs e, AbstractDrawCanvas canvas);
 
-        public virtual void UpdateThickness(int thickness)
+        public virtual void UpdateThickness(double thickness)
         {
             return;
         }
